fix migration

+++ src/database/migrations/2024_06_07_143337_create_table_cashflow.php
@@ -17,7 +17,7 @@ class CreateTableCashflow extends Migration
             $table->unsignedInteger('id', true);
             $table->unsignedSmallInteger('account_id');
             $table->mediumInteger('amount');
-            $table->unsignedBigInteger('created_by');
+            $table->unsignedSmallInteger('created_by');
             $table->string('description', 100);
             $table->timestamps();
