+++ src/database/migrations/2024_06_12_144354_alter_table_purchase_order_add_field_created_by_and_status.php
@@ -15,7 +15,7 @@ class AlterTablePurchaseOrderAddFieldCreatedByAndStatus extends Migration
     {
         Schema::table('purchase_orders', function (Blueprint $table) {
             $table->unsignedInteger('status')->nullable(false)->after('supplier_id');
-            $table->unsignedBigInteger('created_by')->after('supplier_id');
+            $table->unsignedSmallInteger('created_by')->after('supplier_id');
 
             $table->foreign('created_by')->references('id')->on('users');
         });
