+++ src/database/migrations/2024_06_15_042348_alter_table_permission_add_field_is_default.php
@@ -15,7 +15,7 @@ class AlterTablePermissionAddFieldIsDefault extends Migration
     public function up()
     {
         Schema::table('permission', function (Blueprint $table) {
-            $table->unsignedInteger('is_default')->nullable(false)->default(Constant::OPTION_TRUE);
+            $table->unsignedSmallInteger('is_default')->nullable(false)->default(Constant::OPTION_TRUE);
         });
     }
